Fix auto discovery of most recent cylc version
The github api used to find the most recent tagged version of `cylc` is
currently responding with a 301 Moved Permanently response causing
`$CYLC_VERSION` to be empty and the installation of rose and cylc to
abort.

```
{
  "message": "Moved Permanently",
  "url": "https://api.github.com/repositories/1836229/releases",
  "documentation_url": "https://developer.github.com/v3/#http-redirects"
}
```

This commit adds `-L` to the curl commands to ensure that redirects are
followed that the necessary response is given.